Standardize headers

+++ apilib/memcpy.c
@@ -1,6 +1,6 @@
-#include "stdio.h"
+#include "string.h"
 
-void* memcpy(void* dst, const void* src, int size) {
+void* memcpy(void* dst, const void* src, size_t size) {
   char* p = (char*)dst;
   const char* q = (const char*)src;
   for (; size > 0; --size)
