+++ apilib/strncmp.c
@@ -1,6 +1,6 @@
-#include "stdio.h"
+#include "string.h"
 
-int strncmp(const char* s1, const char* s2, int n) {
+int strncmp(const char* s1, const char* s2, size_t n) {
   for (unsigned char *u1 = (unsigned char*)s1, *u2 = (unsigned char*)s2;
        n > 0; ++u1, ++u2, --n) {
     int d = *u1 - *u2;
