+++ apilib/strncpy.c
@@ -1,6 +1,6 @@
-#include "stdio.h"
+#include "string.h"
 
-char* strncpy(char* dst, const char* src, int n) {
+char* strncpy(char* dst, const char* src, size_t n) {
   char* orgDst = dst;
   for (; n > 0; --n) {
     if ((*dst++ = *src++) == '\0')
